Created ApiDownChecker, ApiValidator and HttpValidator
Everything is fully covered by tests.
ApiDownChecker contains the logic to determine if your api is down based on trusted and untrusted validators.
HttpValidator is a handy validator that checks the response of a determined url.
ApiDownChecker.Builder helps construct an instance of ApiDownChecker with proper validators.

We use WireMock for HttpValidator tests that need to go to network.