Adds easier api to create the interceptor
Instead of creating an ApiDownChecker with a builder, and passing it to the ApiInterceptor Builder to create the final interceptor, no we can just call buildInterceptor() on the ApiDownChecker builder directly, and does the boilerplate underneath.
Also added a static create() method to ApiDownChecker to wrap the new builder creation.